fix / refactor: allow pyremote to stop scripts
In this CL, we make some significant changes to pyremote.c so that it can
support *stopping* python scripts during blocking Python calls (e.g.
time.sleep()). First, we introduce the current architecture and main issue.
Then, we explain the design changes to fix this issue.

How pyremote currently works:
- The main Python interpreter exists in the main thread associated to the
plugin.
- Whenever a script is desired to be run (whether starting user script, running
an individual command, or running an action script), a new GTK thread is
spawned, and the Python command(s) called by the interpreter.

Since the 'main' Python interpreter thread is not running Python, it is quite
difficult to properly *stop* a running script. There are two ways to do so:
- PyErr_SetInterruptEx(): this simulates a signal arriving, and is sent to the
Python interpreter *the next time PyErr_CheckSignals() is hit.
- PyErr_SetString(): this sets the *current thread*'s indicator.

Both work when called while the Python interpreter's is blocked, being used
by the Python thread. However, if either is called while the Python thread has
released the GIL, either:
- gxsm crashes, because we are running a Python call without having the GIL; or
- If we grab the GIL, nothing happens (i.e., the Python script does not
stop). This is because the exception state of a Python thread is stored in
the PyThreadState; when the running Python script retakes the GIL, it retakes
the exception state *before* it released it (i.e., before we sent the exception).

     ----------          ----------          ----------          ----------

This refactor modifies the design so we *can* send a kill() signal to the user
script (*not* the action script, as we assume this is the intended design). We
accomplish this by:
- Spawning the main Python interpreter in its own thread, which becomes the
thread for running user scripts/console commands. This is PyRunConsoleThread.
- Spawning a new thread for any Python action script that is spawned. This is
PyRunActionScriptThread.

Since the main interpreter is running the script/commands we may want to stop,
doing so is easy: we simply call PyErr_SetInterruptEx(). If done while the GIL
is released (e.g., during a time.sleep()), the exception is received once the
GIL is retaken.

Actually, I *do not know* why this works; the PyThreadState should contain the
exception state prior to releasing, so this shouldn't be there... Perhaps the
exception state is not reset in the main interpreter thread?

My original implementation for this used Py_AddPendingCall(), to send the signal
the next time the main thread has the GIL. It didn't seem necessary with the
existing design, but we could revisit if we run into issues.

     ----------          ----------          ----------          ----------

The remaining changes are minor refactoring: removing unused attributes I saw,
or whole approaches to this problem that were not working (and wrapped in

Some important notes:
- pyremote_run() was removed, as it is not hit as far as I can see. We set the
plugin attribute that took it to NULL, and saw no ill effects.
- On closing, we currently inform PyRunConsoleThread via a boolean, and wait
for this boolean to be reset (by the thread). This means we could hang forever!
A more proper solution would be to wait a set period of time, checking; if it
has not destroyed itself in that time, we finish the destructor and cause an
ugly exit.
- We are using py_gxsm_console::run_string() to call our strings. This contains
a number of calls sending messages (py_gxsm_console::push_message_async()). I'm
not sure if this is useful or not; we already are printing out all message sent
via Python script into the console via the redirection plugin.
- Related: we are importing both the redirection plugin *and* gxsm both for
user scripts and action scripts. Perhaps we want to *only* import redirection
for the user scripts? This would avoid action script messages showing up in
the Python console. If desired, one would simply add a boolean to
py_gxsm_console::create_environment() for this.
- We are currently keeping the static struct PyGxsmModuleInfo, even though it
only has 1 used PyObject! This could easily be pulled into py_gxsm_console, but
I did not know if this was appropriate.