feat: kill python 2 different ways to be sure
Killing the interpreter with an interrupt via Py_AddPendingCall() does
not appear to work if we are using gxsm.sleep(). This is because the
main thread never gets called, because we are never releasing the
GIL. (I am not sure *why* we do not release the GIL, since the default
switch interval is not none...). However, this approach *does* work with
more standard Python calls, where we release the GIL to do I/O work or
sleep.

To account for gxsm.sleep(), we add a sleep after calling
Py_AddPendingCall(), and throw a PyErr_SetInterruptEx (SIGINT). This
works for gxsm.sleep(), but not time.sleep().

The downside to this approach: if we exit due to gxsm.sleep(), we end up
with an *extra* excpetion in the error stack, from the original
Py_AddPendingCall() call. This means that the *next* time you run a
script (after crashing with gxsm.sleep()) will fail. This could be
remedied with a call to PyErr_Clear(); I will investigate this further
if this resolves issues.