fix: Better forcing of exception
The prior approach did not work well, because 2 exceptions are actually
always sent! this means that, even though our exception is sent, we do
not give our Python code the ability to properly react to our raised
exception.

It appears that a more 'proper' workaround is to try and 'force' Stop()
to be called (even if the GIL is not being released), by explicitly
requesting the GIL in our kill() call. Note that PyErr_CheckSignals()
may not even be necessary.