fix: allow python actions+main script to run
Allow python actions to run simultaneously with the main script, without
stopping each other. Previously, when one ended it was causing the other
to end as well!

After investigation, it was found that this was due to PyDict_Clear(d)
in destroy_environment(). My best explanation: PyDict_Copy() is a
shallow, not deep, copy. Since there is some mutable parameter in the
__main__ dict, when we call PyDict_Clear(), that gets cleared. Whatever
mutable parameter this is, it tells the interpreter to stop running!

To work around this, I add an optional (default to false) clearGlobals
variable to destroy_environment(). Note that we are still decreasing our
dict references appropriately, for most cases. We only call clearGlobals
when the main python thread is actually ending.