feat: add menu option restart python interpreter
Here, we add a menu item to pyremote, to restart the python
interpreter, and modify py_gxsm_console to do so. What it does is tells
us to call destroy_environment() with clearGlobals = true.

I don't fully understand what is going on, but something in destroying
the main pydict forces a restart of all variables in the interpreter (as
far as I can tell). This is why we stopped fully destroying the dict by
default. Now, it only happens when (a) PyRunConsoleThread
closes (presumably when gxsm ends), or (b) when we call this menu item.

In short: it is possible this is not a proper *restart*! If so, we will
need to modify what restart_interpreter_callback() does.